backend/requirements.txt
greenlet~=1.1.3
werkzeug~=2.0.0
opencv-python
numpy
face_recognition
flask
flask_restful
flask_cors
flask-reuploaded
flask_sqlalchemy
pillow
dlib
sqlalchemy~=1.4.48
cmake
thread
marshmallow~=3.19.0
marshmallow_sqlalchemy
python-decouple
imutils~=0.5.4
python-dotenv
flask_jwt_extended
